refactor(자동차 객체 변수) : 자동차 객체의 변수 Long형으로 변경
무한한 수를 받을 수 있는 BigInteger에서 메모리, 성능 문제를 고려해서 Long형으로 변경

+++ src/main/java/racingcar/model/RacingCar.java
@@ -1,6 +1,6 @@
 package racingcar.model;
 
-import java.math.BigInteger;
+
 import java.util.Objects;
 import racingcar.utils.validator.CarValidator;
 
@@ -8,9 +8,9 @@ public class RacingCar implements Car{
 
     private final String name;
 
-    private final BigInteger progress;
+    private final Long progress;
 
-    public RacingCar(final String name, final BigInteger progress) {
+    public RacingCar(final String name, final Long progress) {
         this.validate(name);
 
         this.name = name;
@@ -26,7 +26,7 @@ public class RacingCar implements Car{
         return name;
     }
 
-    public BigInteger getProgress() {
+    public Long getProgress() {
         return progress;
     }
